+++ packages/router/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Rename `route.response` to `route.respond`.
 * `meta`, `body`, and `data` properties always exist on response objects. They are `undefined` if not set by a route's `response()` function.
 * Add `meta` property to response/settable response properties.
 * Remove `title`, `status`, and `error` types from response/settable response properties.